# backend/docs/api/market-tracker.openapi.yaml
openapi: 3.1.0
info:
  title: Market Tracker API
  version: 0.1.0-draft
  description: |
    Contract-first REST API draft derived from:
    - backend/docs/Market Tracker.md
    - backend/docs/DB_SCHEMA.md
    - backend/docs/EVENT_LOGIC.md
    - backend/docs/BE_ARCHITECTURE.md

    Design assumptions:
    - All routes are scoped by `workspace_id`.
    - Public identifiers use `tracker_code`, `job_code`, `event_code`, and `digest_code`.
    - Responses are built from normalized read models, not raw provider payloads.
    - Category tracking is fixed to Top 50, with Top 10 alerting as a secondary signal.
servers:
  - url: https://api.example.com
    description: Placeholder server
...
paths:
...
  /v1/workspaces/{workspace_id}/events:
    get:
      tags: [events]
      operationId: listEvents
      summary: List events with filters
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdPath'
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
        - $ref: '#/components/parameters/FromDateQuery'
        - $ref: '#/components/parameters/ToDateQuery'
        - name: tracker_type
          in: query
          required: false
          schema:
            type: string
            enum: [CATEGORY, COMPETITOR]
        - name: tracker_code
          in: query
          required: false
          schema:
            type: string
        - name: marketplace
          in: query
          required: false
          schema:
            type: string
            pattern: '^amazon_[a-z]{2}$'
        - name: asin
          in: query
          required: false
          schema:
            type: string
            minLength: 10
            maxLength: 10
        - name: event_type
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/EventType'
        - $ref: '#/components/parameters/SeverityQuery'
      responses:
        '200':
          description: Paged event list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /v1/workspaces/{workspace_id}/reports/weekly-digests:
    get:
      tags: [reports]
      operationId: listWeeklyDigests
      summary: List weekly digests
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdPath'
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
        - name: week_start
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Paged weekly digest list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyDigestListResponse'
  /v1/workspaces/{workspace_id}/reports/weekly-digests/{digest_code}:
    get:
      tags: [reports]
      operationId: getWeeklyDigest
      summary: Get weekly digest detail
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdPath'
        - $ref: '#/components/parameters/DigestCodePath'
      responses:
        '200':
          description: Weekly digest detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyDigest'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  parameters:
    WorkspaceIdPath:
      name: workspace_id
      in: path
      required: true
      schema:
        type: string
      description: Stable workspace identifier.
...
    DigestCodePath:
      name: digest_code
      in: path
      required: true
      schema:
        type: string
...
    PageQuery:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeQuery:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
    FromDateQuery:
      name: from_date
      in: query
      required: false
      schema:
        type: string
        format: date
    ToDateQuery:
      name: to_date
      in: query
      required: false
      schema:
        type: string
        format: date
...
    SeverityQuery:
      name: severity
      in: query
      required: false
      schema:
        type: string
        enum: [LOW, MEDIUM, HIGH]
...
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: VALIDATION_ERROR
            message: Invalid request body.
            details:
              field: scope.browse_node_id
              reason: Either browse_node_id or browse_node_url must be provided.
            request_id: req_01J8X9XYZ
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: NOT_FOUND
            message: Resource not found.
            request_id: req_01J8X9XYZ
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        request_id:
          type: string
    Marketplace:
      type: string
      pattern: '^amazon_[a-z]{2}$'
      examples: [amazon_us, amazon_de]
...
    EventType:
      type: string
      enum:
        - NEW_ENTRANT_TOP50
        - RETURNING_TOP50
        - EXIT_TOP50
        - ENTER_TOP10
        - EXIT_TOP10
        - PRICE_CHANGED
        - PROMOTION_CHANGED
        - TITLE_CHANGED
        - MAIN_IMAGE_CHANGED
        - VARIATIONS_ADDED
        - CONTENT_CHANGED
        - AVAILABILITY_CHANGED
        - BUY_BOX_CHANGED
...
    AvailabilityStatus:
      type: string
      enum: [IN_STOCK, OUT_OF_STOCK, INACTIVE, UNKNOWN]
    BuyBoxStatus:
      type: string
      enum: [HAS_BUY_BOX, NO_BUY_BOX, UNKNOWN]
    EventListResponse:
      type: object
      required: [items, page, page_size, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
    Event:
      type: object
      required:
        - event_code
        - tracker_type
        - tracker_code
        - marketplace
        - asin
        - event_type
        - event_time
        - snapshot_date
        - severity
        - title
        - summary
        - payload
      properties:
        event_code:
          type: string
        tracker_type:
          type: string
          enum: [CATEGORY, COMPETITOR]
        tracker_code:
          type: string
        marketplace:
          $ref: '#/components/schemas/Marketplace'
        asin:
          type: string
        event_type:
          $ref: '#/components/schemas/EventType'
        event_time:
          type: string
          format: date-time
        snapshot_date:
          type: string
          format: date
        severity:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        title:
          type: string
        summary:
          type: string
        payload:
          $ref: '#/components/schemas/EventPayload'
        job_code:
          type: string
        dedupe_key:
          type: string
    EventPayload:
      type: object
      additionalProperties: true
      properties:
        rank_today:
          type: integer
        first_seen_in_tracker:
          type: boolean
        last_seen_date:
          type: string
          format: date
        days_absent:
          type: integer
        previous_rank:
          type: integer
        current_rank:
          type: integer
        present_today:
          type: boolean
        previous:
          $ref: '#/components/schemas/EventChangeState'
        current:
          $ref: '#/components/schemas/EventChangeState'
        delta:
          $ref: '#/components/schemas/EventDelta'
    EventChangeState:
      type: object
      additionalProperties: true
      properties:
        price_current:
          type: number
          format: float
          nullable: true
        price_original:
          type: number
          format: float
          nullable: true
        coupon_text:
          type: string
          nullable: true
        title:
          type: string
        main_image_url:
          type: string
          format: uri
        variation_count:
          type: integer
        content_signature_hash:
          type: string
        a_plus_signature_hash:
          type: string
        availability_status:
          $ref: '#/components/schemas/AvailabilityStatus'
        buy_box_status:
          $ref: '#/components/schemas/BuyBoxStatus'
        buy_box_seller_name:
          type: string
    EventDelta:
      type: object
      properties:
        price_current_abs:
          type: number
          format: float
        price_current_pct:
          type: number
          format: float
...
    TrackerRef:
      type: object
      required: [tracker_type, tracker_code, tracker_name]
      properties:
        tracker_type:
          type: string
          enum: [CATEGORY, COMPETITOR]
        tracker_code:
          type: string
        tracker_name:
          type: string
...
    WeeklyDigestListResponse:
      type: object
      required: [items, page, page_size, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/WeeklyDigest'
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
    WeeklyDigest:
      type: object
      required:
        - digest_code
        - week_start
        - week_end
        - tracker_refs
        - summary
        - threats
        - created_at
      properties:
        digest_code:
          type: string
        week_start:
          type: string
          format: date
        week_end:
          type: string
          format: date
        tracker_refs:
          type: array
          items:
            $ref: '#/components/schemas/TrackerRef'
        summary:
          $ref: '#/components/schemas/WeeklyDigestSummary'
        threats:
          type: array
          items:
            $ref: '#/components/schemas/Threat'
        report_storage_uri:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    WeeklyDigestSummary:
      type: object
      required:
        - new_entrant_count
        - returning_count
        - top10_enter_count
        - price_change_count
        - listing_change_count
      properties:
        new_entrant_count:
          type: integer
        returning_count:
          type: integer
        top10_enter_count:
          type: integer
        price_change_count:
          type: integer
        listing_change_count:
          type: integer
    Threat:
      type: object
      required: [asin, marketplace, reason, tracker_refs]
      properties:
        asin:
          type: string
        marketplace:
          $ref: '#/components/schemas/Marketplace'
        reason:
          type: string
        event_types:
          type: array
          items:
            $ref: '#/components/schemas/EventType'
        tracker_refs:
          type: array
          items:
            $ref: '#/components/schemas/TrackerRef'